+++ src/libs/highlevel/CMakeLists.txt
@@ -1,3 +1,3 @@
 file (GLOB SOURCES
 	   *.c)
-add_lib (highlevel SOURCES ${SOURCES} elektra_conversion_doxygen.h LINK_ELEKTRA elektra-core elektra-kdb elektra-ease)
+add_lib (highlevel SOURCES ${SOURCES} LINK_ELEKTRA elektra-core elektra-kdb elektra-ease)
